requirements.txt
asgiref==3.9.1
Django==5.2.6
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==11.3.0
PyMySQL==1.1.2
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
